+++ frontend/src/redux/ducks/auth/auth.slice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { IFetchedLogin, IFetchedRegistration, ILoginFailed } from "./action.types";
+import { IFetchedLogin, IFetchedRegistration, ILoginFailed, IRegistrationFailed } from "./action.types";
 import { ILoginResponse, ISuccessResponse } from "../../../api/auth.api";
 
 interface AuthState {
@@ -16,35 +16,42 @@ const initialState: AuthState = {
     loading: false,
     error: null,
     token: null
-}
+};
 
 export const authSlice = createSlice({
-    name: 'auth',
+    name: "auth",
     initialState,
     reducers: {
-        fetchedLogin(state, action: PayloadAction<IFetchedLogin>) {
-            state.loading = true
-        },
-        loginSuccess(state, action: PayloadAction<ILoginResponse>) {
-            state.loading = false
-            state.token = action.payload.token
-        },
-        loginFailed(state, action: PayloadAction<ILoginFailed>) {
-            state.loading = false
-            state.error = action.payload.error
-        },
         fetchedRegistration(state, action: PayloadAction<IFetchedRegistration>) {
-            state.loading = true
+            state.loading = true;
         },
         registrationSuccess(state, action: PayloadAction<void>) {
-            state.loading = false
+            state.loading = false;
         },
-        registrationFailed(state, action: PayloadAction<ILoginFailed>) {
-            state.loading = false
-            state.error = action.payload.error
+        registrationFailed(state, action: PayloadAction<IRegistrationFailed>) {
+            state.loading = false;
+            state.error = action.payload.error;
+        },
+        fetchedLogin(state, action: PayloadAction<IFetchedLogin>) {
+            state.loading = true;
+        },
+        loginSuccess(state, action: PayloadAction<ILoginResponse>) {
+            state.loading = false;
+            state.token = action.payload.token;
+        },
+        loginFailed(state, action: PayloadAction<ILoginFailed>) {
+            state.loading = false;
+            state.error = action.payload.error;
         }
     }
-})
+});
 
-export const { fetchedLogin, loginSuccess, loginFailed, fetchedRegistration, registrationSuccess, registrationFailed } = authSlice.actions
-export const authReducer = authSlice.reducer
+export const {
+    fetchedLogin,
+    loginSuccess,
+    loginFailed,
+    fetchedRegistration,
+    registrationSuccess,
+    registrationFailed
+} = authSlice.actions;
+export const authReducer = authSlice.reducer;