+++ frontend/src/types/profile/unfollow.dto.ts
@@ -1,0 +1,5 @@
+export interface UnfollowDto {
+    user_id: number;
+    unfollow_id: number;
+
+}