+++ frontend/src/types/profile/updateAvatar.dto.ts
@@ -1,0 +1,4 @@
+export interface UpdateAvatarDto {
+    id: number;
+    avatar: File;
+}