+++ frontend/src/types/profile/updateStatus.dto.ts
@@ -1,0 +1,5 @@
+
+export interface UpdateStatusDto {
+    id: number;
+    status: string
+}